+++ src/main/java/com/gukbit/api/controller/CourseController.java
@@ -1,4 +1,4 @@
-package com.gukbit.controller;
+package com.gukbit.api.controller;
 
 import com.gukbit.domain.Course;
 import com.gukbit.service.CourseService;
