+++ src/main/java/com/gukbit/api/controller/IndexController.java
@@ -1,4 +1,4 @@
-package com.gukbit.controller;
+package com.gukbit.api.controller;
 
 
 import com.gukbit.domain.Course;
